+++ v21_gfp_exact_sched_test/2processors/generate_successors.h
@@ -1,1 +1,1 @@
-void generate_successors(const TS&, state&, const uint_fast8_t, vector<state>*, uint_fast16_t&);
+void generate_successors(const TS&, state&, const unsigned short, vector<state>*, unsigned int&);
