+++ v21_gfp_exact_sched_test/2processors/response_time_Guan2009.h
@@ -1,2 +1,2 @@
-uint_fast8_t* response_time_Guan2009(const TS& ts, const uint_fast8_t m);
+unsigned short* response_time_Guan2009(const TS& ts, const unsigned short m);
